requirements.txt
crewai
crewai-tools
langchain
langchain-community
langchain-openai
langchain-groq
langchain-anthropic
langchain-ollama
streamlit
python-dotenv
pdfminer.six
sqlalchemy
psycopg2-binary
snowflake-connector-python
markdown
docling
duckduckgo-search>=8.0.2
boto3
botocore
cryptography
pandas
snowflake-snowpark-python
crewAI-tools
